Add SimplePrefixNamingRule class
