+++ Samples/Editor/SimplePrefixNamingRule.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: c008d0e670354828842750c400d6d075
+timeCreated: 1628865848
